config: reduce default topic_fds_per_partition
Now that segment_reader and segment_index open files
on demand, we can be a bit less paranoid about the number
of FDs required per partition.

Set the default to 6:
- 3 for an ongoing compaction (reading two segments and writing 1)
- 1 for reading the latest segment
- 1 for writing the latest segment

In practice most topics will use fewer FDs at a given moment
than this, but this new limit is more aligned with real sources
of FD consumption than the previous arbitrary 10.