tests: fix test_configuration_properties_kafka_config_allowlist
This test was previously fetching a TopicSpec for an existing
topic and using that TopicSpec's segment_bytes to derive a
different value for a config change.

Except, that describe operation was never populating segment_bytes,
they were just getting the default on the TopicSpec class.  That
recently changed to None, which highlighted the bad behaviour.

Modify the test to use rpk.describe_topic_configs instead, which
does fetch the segment_bytes correctly.