storage: handle ENOENT in maybe_disable_cow
In opfuzz, there can be situations where a file is
deleted around the same time as it is opened elsewhere: this
leads to the file no longer being present by the time
maybe_disable_cow tries to stat it.

maybe_disable_cow is nonessential, so it should just ignore
this error: whoever opened the file still gets their
file handle, unlinked though it may be.