Use random tie-breaking in partition allocator
Currently we always select the node with the lowest ID out of nodes
with equal scores when allocating new partitions. This is largely a
side effect of the scan for the lowest score, but it means that
replica sets chosen tend to be highly concentrated: only a few replica
sets are chosen in practice even if many are available.

For example, in a cluster with 6 nodes, when the partition is balanced,
new partitions will be assigned only to replica sets [0 1 2] or [3 4 5]
despite there being 18 other valid replica sets to choose from.

This change breaks ties randomly, resulting in (probabilistically) a
relatively even spread of chosen replica sets.

Fixes #4809.