k/groups: Release the ssg semaphore for 2nd stage
Currently we hold the kafka smp_service_group semaphore (by using
invoke_on with that ssg) for the duration of group protocol requests
such as JoinGroup. This means that the maximum number of outstanding
requests to JoinGroup is limited to the ssg limit. This means that
the number of consumers which can join a group has the same limit
as all members must reach the joining state before the join completes.

See comment in group_router::route_stages for additional details.

Fixes #4327.