hold mem units until the response is written out
Currently we dropped the units just before passing the scattered
buffer to the connection to be written (because the response
object is destroyed at this point), which leaves a window where
the buffer for the response still exists in-memory but the memory
units have already been released back to the semaphore.

With the right combination of events, this can result in an OOM event.