Improved (?) comment for for fetch/produce latency
We have a fetch and produce latency metric which sort of source like
a fetch/produce request latency metric, but in fact it measures the
internal latency of a produce to or fetch from a single partition,
and not the entire request. This can be much smaller than the request
latency because a request may have many partitions and also because it
includes large parts of the request entirely such as reading the
request off the wire and returning the result.