Remove net::server::hist
net::server has an embedded histogram object _hist, but never itself
updated the histogram, but relied on the subclasses (of which there
are only two: the kafka and rpc server) to udpate it.

The two subclasses did it in different ways, and kafka specifically
did something a bit special: it only adds produce + fetch times to
the histogram. This is mostly useless since fetch may have very large
latency due to polling and per-handler metrics offers a superset of
functionality.

This change implements the old kafka histo just by adding together the
existing produce + fetch per handler histos, and moves the rpc
histogram to the subclass.