Implement async_for_each
Same as std::for_each, but which yields periodically. Written in a
style which keeps in inner loop fast, equivalent to the non-async
version (optimizer willing, but seems to be true in the case I
checked).

For the simplest use cases we offer async_for_each, which
is the same as std::for_each except that it yields every
Traits::Interval iterations: every call starts the "work counter" anew.

For more complicated cases, like nested loops, it may be convenient to
carry the counter from one iteration to the next, in order to yield at
the correct times. For example, in a doubly nested loop where the inner
loop always ran for less than Interval iterations, we would not yield
with the simple functions.

Issue redpanda-data/core-internal#1061.